arm64: Add cou hotplug for spin table method
This patch has been implemented for CPU hotplug support for spin-table
method.

While cpu is offlined, cpu_die is called and when it is brought online
cpu_boot is called. So, cpu_boot must wake secondary and release pen,
otherwise dynamic cpu offline/online will not work.

Upstream is in favor of psci method, so use it until firmware with psci
support is available.

Signed-off-by: Pratyush Anand <panand@redhat.com>
Signed-off-by: Matheus Castello <matheus@castello.eng.br>