arm64: kexec: Fix the crash kernel call
This works for Pi 3B +, I don't care about the status of the registers.
I want to be able to load a new kernel and get the vmcore dmesg, with
these modifications I can do this. It is a workaround for the method
without psci so if you want to use kdump in your project with Pi 3B+
think about the objective you want to achieve, this is not the right
way to do this.

Signed-off-by: Matheus Castello <matheus@castello.eng.br>