soc: sunxi: sram: Add support for the D1 system control
D1 has a single EMAC and some LDOs that need to be exported.

Cover-letter:
soc: sunxi: sram: Fixes and D1 support
This series cleans up a few issues in the system controller driver, and
then expands the exported regmap to support one of the pairs of LDOs
built in to the D1 SoC.

Eventually, we will need to update the SRAM region claiming API so
ownership can be swapped back and forth by consumer drivers. This is
necessary for uploading firmware to the R329/D1 DSPs, but it is not
needed for initial bringup.
END

Series-to: Chen-Yu Tsai <wens@csie.org>
Series-to: Jernej Skrabec <jernej.skrabec@gmail.com>

Signed-off-by: Samuel Holland <samuel@sholland.org>