dt-bindings: crypto: sun8i-ce: Add compatible for D1
D1 has a crypto engine similar to the one in other Allwinner SoCs.
Like H6, it has a separate MBUS clock gate.

It also requires the internal RC oscillator to be enabled for the TRNG
to return data. This is likely the case for earlier variants as well,
but the clock drivers for earlier SoCs did not allow disabling the RC
oscillator.

Series-changes: 2
 - Add TRNG clock

Signed-off-by: Samuel Holland <samuel@sholland.org>