rtc: sun6i: add support for R329 RTC
Allwinner R329 has a RTC with a similar time storage with H616 but a
slightly different clock part.

As we have already handled the R329 RTC clocks in the CCU driver, add a
compatible string to RTC driver to allow probing of the RTC.

Commit-notes:
Resending this patch separately from Icenowy's R329 series[1] because it
is also needed for D1 (which has R329 as its fallback compatible[2]), so
I would like to get it in to 5.20.

[1]: https://lore.kernel.org/lkml/BYAPR20MB2472C608678F3FAEDA7B7541BCF79@BYAPR20MB2472.namprd20.prod.outlook.com/
[2]: https://lore.kernel.org/lkml/20220203021736.13434-3-samuel@sholland.org/
END

Series-prefix: RESEND
Series-to: Alessandro Zummo <a.zummo@towertech.it>
Series-to: Alexandre Belloni <alexandre.belloni@bootlin.com>

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>
Signed-off-by: Samuel Holland <samuel@sholland.org>