mfd: sun6i-prcm: Update Kconfig description
This driver is used by a specific range of SoCs between when the PRCM
hardware was introduced (A31) and when the PRCM CCU driver was added
(A83T). It is unlikely to be extended to additional hardware. Update
the description to include the full list of applicable SoCs.

Series-to: Lee Jones <lee.jones@linaro.org>
Series-cc: Chen-Yu Tsai <wens@csie.org>
Series-cc: Jernej Skrabec <jernej.skrabec@gmail.com>
Series-cc: linux-sunxi@lists.linux.dev

Signed-off-by: Samuel Holland <samuel@sholland.org>