soc: sunxi: sram: Only iterate over SRAM children
Now that a "regulators" child is accepted by the controller binding, the
debugfs show routine must be explicitly limited to "sram" children.
Otherwise, it will crash because the "regulators" node has no address.

Series-to: Liam Girdwood <lgirdwood@gmail.com>
Series-to: Mark Brown <broonie@kernel.org>
Series-to: Chen-Yu Tsai <wens@csie.org>
Series-to: Jernej Skrabec <jernej.skrabec@gmail.com>
Series-to: Krzysztof Kozlowski <krzysztof.kozlowski+dt@linaro.org>
Series-to: Rob Herring <robh+dt@kernel.org>

Commit-changes: 2
 - New patch for v2

Series-version: 3

Cover-letter:
regulator: Add support for Allwinner D1 LDOs
This series adds bindings and a driver for the two pairs of LDOs
inside the Allwinner D1 SoC.

A preparatory binding and driver change is required for the SRAM
controller, so the regulators device can be its child node.
END

Signed-off-by: Samuel Holland <samuel@sholland.org>